sdk: move DarwinSDK into PackLib
