+++ Sources/XToolSupport/SetupCommand.swift
@@ -1,6 +1,7 @@
 import ArgumentParser
 import Foundation
 import XKit
+import PackLib
 
 struct SetupCommand: AsyncParsableCommand {
     static let configuration = CommandConfiguration(
